### 2017-02-25 #186 @whedon
[PRE REVIEW]: MSMExplorer: Data Visualizations for Biomolecular Dynamics
**Submitting author:** @cxhernandez (<a href="http://orcid.org/0000-0002-8146-5904">Carlos Xavier Hernández</a>)
**Repository:** <a href="https://github.com/msmexplorer/msmexplorer" target ="_blank">https://github.com/msmexplorer/msmexplorer</a>
**Version:** v1.0.0
