### 2017-02-25 #186 @arokem
Actually, I'd be happy to edit this! And I have a potential reviewer in mind...

@anotherjoshsmith, would you like to review this paper for JOSS (see: http://joss.theoj.org/about#reviewer_guidelines)?
