### 2017-03-02 #188 @arokem
@arfon : why does this say "Does the release version given match the GitHub release (v1.0.0)?"?

The current release version on Github is 0.3. Where does the `1.0.0` come from?
