### 2017-03-02 #188 @cxhernandez
@arokem: That's my fault. My intention was to hold off on making a v1.0.0 release until the review is done. In my previous experience with JOSS, the reviewer at the time requested additional functionality to package, and it required us to cut a new release anyways.
