### 2017-03-02 #188 @arokem
Gotcha - thanks! That means that @anotherjoshsmith should go through everything else, wait until everything else is in order, and then wait for you to cut a v1.0.0 before checking that box. Is that your intention?
