### 2017-03-03 480768a
minor free energy fix
