+++ msmexplorer/plots/projection.py
@@ -132,7 +132,7 @@ def plot_free_energy(data, ax=None, obs=0, temperature=300., n_samples=None,
 
         if shade:
             ax.fill_between(X, Z - Z.min(), Z.max() - Z.min(),
-                        facecolor=color, alpha=alpha)
+                            facecolor=color, alpha=alpha)
 
     elif prune.shape[1] == 2:
 
