### 2017-03-09 msmexplorer/msmexplorer#77 @anotherjoshsmith
[JOSS REVIEW]
_The following tasks should be completed for JOSS acceptance:_

__Installation:__
- [x] Support conda install with python 3.6.
- [x] Specify version requirements (if applicable) for dependencies. The Fs-peptide notebook gets angry at pandas 0.16 (a few years behind the times) for the tICA cell. Alternatively, suggest that users update to most recent version for these dependencies if they are using an existing python 3.4 build.

__Example usage:__
- [x] Describe a possible use case for the chord diagram. It seems a bit disconnected from the rest of the tools, so adding an example (perhaps to some visualization of FS-peptide structure) could really tie things together.

__Community guidelines:__
- [x] Add a sentence to the "Development" section of your README directing potential contributors to the link you have already included.
- [x] Include a preferred method of contact for users seeking support.

__Version:__
- [x] Cut v1.0.0
