### 2017-03-10 msmexplorer/msmexplorer#77 @cxhernandez
Thanks for the review, @anotherjoshsmith!

> Support conda install with python 3.6.

I think it should already be supported [under the omnia channel](https://anaconda.org/omnia/msmexplorer/files), but I've also been meaning to add msmexplorer to conda-forge as well.

As for the other items, I'll get started on a PR to address them.
