### 2017-03-10 msmexplorer/msmexplorer#77 @anotherjoshsmith
`conda install -c omnia msmexplorer` doesn't work for me on a fresh python 3.6 build on OSX, but everything installs fine with python 3.4. I'm having the same problem with msmbuilder on 3.6.
