### 2017-03-10 msmexplorer/msmexplorer#77 @Lnaden
The problem is that the `msmbuilder` was uploaded under the `rc` tag and not the `main` tag, so a standard install procedure will miss it. Let me tag the author of that package to see why its on the rc tag and see if I can get a solution for you.
