### 2017-03-10 msmexplorer/msmexplorer#77 @Lnaden
If you need a "right-now-right-now" solution, you can add the `-c omnia/label/rc` to your `conda install` and and it *should*(?) install
