### 2017-03-23 7f365bb
added option to return the data that is used to plot
