+++ msmexplorer/plots/projection.py
@@ -27,7 +27,7 @@ def plot_free_energy(data, ax=None, obs=0, temperature=300., n_samples=None,
                      vmin=None, vmax=None, n_levels=10, clabel=False,
                      clabel_kwargs=None, cbar=False, cbar_kwargs=None,
                      xlabel=None, ylabel=None,
-                     labelsize=14, random_state=None):
+                     labelsize=14, random_state=None, return_data = False):
     """
     Plot free energy of observable(s) in kilocalories per mole.
 
@@ -92,12 +92,17 @@ def plot_free_energy(data, ax=None, obs=0, temperature=300., n_samples=None,
         The generator used to initialize the centers. If an integer is
         given, it fixes the seed. Defaults to the global numpy random
         number generator
+    return_data : Boolean,optional
+        Whether or not to return the plotting data
 
     Returns
     -------
     ax : matplotlib axis
         matplotlib figure axis
-
+    return_data : list of lists, optional
+        data that can be used to remake the figures. The last list is always the free energy.
+        To remake 1 dim plot:plot(return_data[0],return_data[1]).
+        To remake 2 dim plots: contour(return_data[0],return_data[1],return_data[2])
     """
 
     sns.set_style('whitegrid')
@@ -119,6 +124,9 @@ def plot_free_energy(data, ax=None, obs=0, temperature=300., n_samples=None,
         idx = random_state.choice(range(data.shape[0]), size=n_samples, p=pi)
         prune = prune[idx, :]
 
+    if return_data:
+        to_return = []
+
     if prune.shape[1] == 1:
 
         if clip is None:
@@ -133,6 +141,9 @@ def plot_free_energy(data, ax=None, obs=0, temperature=300., n_samples=None,
         if shade:
             ax.fill_between(X, Z - Z.min(), Z.max() - Z.min(),
                             facecolor=color, alpha=alpha)
+        if return_data:
+            to_return.append(X)
+            to_return.append(Z)
 
     elif prune.shape[1] == 2:
 
@@ -176,6 +187,10 @@ def plot_free_energy(data, ax=None, obs=0, temperature=300., n_samples=None,
 
         ax.grid(zorder=0)
 
+        if return_data:
+            to_return.append(X)
+            to_return.append(Y)
+            to_return.append(Z)
     else:
         raise ValueError('obs cannot be greater than size 2')
 
@@ -185,4 +200,7 @@ def plot_free_energy(data, ax=None, obs=0, temperature=300., n_samples=None,
     if ylabel:
         ax.set_ylabel(ylabel, size=labelsize)
 
-    return ax
+    if return_data:
+        return ax, to_return
+    else:
+        return ax
