+++ examples/plot_free_energy_1d.py
@@ -2,6 +2,7 @@
 Free Energy Plot (Univariate)
 =============================
 """
+from msmbuilder.example_datasets import FsPeptide
 from msmbuilder.featurizer import DihedralFeaturizer
 from msmbuilder.decomposition import tICA
 from msmbuilder.cluster import MiniBatchKMeans
@@ -10,7 +11,6 @@ from msmbuilder.msm import MarkovStateModel
 import numpy as np
 
 import msmexplorer as msme
-from msmexplorer.example_datasets import FsPeptide
 
 rs = np.random.RandomState(42)
 
