### 2017-03-30 msmexplorer/msmexplorer#77 @cxhernandez
@anotherjoshsmith: Please review the changes made in #79, when you get a chance.

I updated the [chord diagram example](https://github.com/cxhernandez/msmexplorer/blob/6e7273d4ed5c33a921ebea12a4c79d01e905f9c0/examples/plot_chord.py) to show how it can be used to plot a hydrogen bonding network in Fs Peptide:

![image](https://cloud.githubusercontent.com/assets/5142365/24494138/1a36b552-14e6-11e7-9b26-6667685651e8.png)
