+++ docs/papers.rst
@@ -6,3 +6,22 @@ Publications Using MSMExplorer
 Below is an ever-growing list of publications that use MSMExplorer plots. If
 you've used MSMExplorer in your paper, feel free to add it here via pull
 request!
+
+
+A Network of Conformational Transitions in the Apo Form of NDM-1 Enzyme Revealed by MD Simulation and a Markov State Model
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. image:: http://pubs.acs.org/appl/literatum/publisher/achs/journals/content/jpcbfk/0/jpcbfk.ahead-of-print/acs.jpcb.7b00062/20170329/images/medium/jp-2017-00062b_0008.gif
+
+.. code:: bibtex
+
+  @article{Gao2017,
+    doi = {10.1021/acs.jpcb.7b00062},
+    url = {https://doi.org/10.1021%2Facs.jpcb.7b00062},
+    year  = {2017},
+    month = {mar},
+    publisher = {American Chemical Society ({ACS})},
+    author = {Kaifu Gao and Yunjie Zhao},
+    title = {A Network of Conformational Transitions in the Apo Form of {NDM}-1 Enzyme Revealed by {MD} Simulation and a Markov State Model},
+    journal = {The Journal of Physical Chemistry B}
+  }
