### 2017-04-01 #188 @cxhernandez
> I think you have put together a very useful research tool. I've even started testing it with data sets of my own!

That's great to hear! Thanks!

I've made a [PR](https://github.com/msmexplorer/msmexplorer/pull/79) that addresses each task item. Let me know what you think!
