### 2017-04-02 msmexplorer/msmexplorer#77 @anotherjoshsmith
@cxhernandez Looks great! Chord diagram example is just what I had in mind and issues with python 3.6 installation seem to be fixed.
