### 2017-04-04 317d4b6
Update changelog.rst
