### 2017-04-04 msmexplorer/msmexplorer#77 @cxhernandez
@anotherjoshsmith: PR has been merged and v1.0.0 should be out in the wild.
