### 2017-04-05 #188 @anotherjoshsmith
All [review issues](https://github.com/msmexplorer/msmexplorer/issues/77#issuecomment-291564733) have been resolved.

@arokem: I suggest _"MSMExplorer: Data Visualizations for Biomolecular Dynamics"_ by @cxhernandez et al. be accepted by JOSS.
