### 2017-04-06 439d997
Add py36 to tests
