### 2017-04-07 #188 @arokem
Thank you @anotherjoshsmith!

@cxhernandez : could you please create a Zenodo archive with the 1.0.0 version? Once you do that, I can set that as the archive, and move ahead to accept and close this issue.
