### 2017-04-07 #188 @cxhernandez
@arokem: The archive for v1.0.0 can be found [here](https://doi.org/10.5281/zenodo.439774).
