### 2017-04-08 #188 @arfon
> @arfon : I believe this one is ready to go! Anything else I need to do to get this one into this page: http://joss.theoj.org/papers/popular?

@arokem - I still have to do the last few bits. Planning on giving @whedon those super-powers sometime soon.

@arokem many thanks for editing this and @anotherjoshsmith for your review ✨

@cxhernandez - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00188 :zap: 🚀 💥
